requirements.txt
numpy
scipy
matplotlib
pandas
opencv-python
Pillow
imageio
psutil
plotly
statsmodels
moepy